+++ src/application/application.ts
@@ -27,12 +27,22 @@ class Application {
         await this.startService(event.serviceID);
         await this.startService(task.serviceID);
 
+        if (typeof event.filter === 'string') {
+            console.warn("Please use eventKey in order to filter on a specific event");
+            event.eventKey = event.filter;
+            event.filter = null;
+        }
+        event.filter = event.filter || ((eventKey, eventData) => true);
+
         const stream = this.client.listenEvent({
             serviceID: event.serviceID,
-            eventFilter: event.filter || '*'
+            eventFilter: event.eventKey || '*'
         })
         stream.on('data', async ({ eventKey, eventData }) => {
-            await this.executeTask(task, eventKey, eventData);
+            const filter = event.filter as (eventKey: string, eventData: Object) => boolean;
+            if (filter(eventKey, JSON.parse(eventData))) {
+                await this.executeTask(task, eventKey, eventData);
+            }
         });
         return stream;
     }
@@ -41,20 +51,24 @@ class Application {
         await this.startService(result.serviceID);
         await this.startService(task.serviceID);
 
+        result.filter = result.filter || ((outputKey, outputData) => true);
+        
         const stream = this.client.listenResult({
             serviceID: result.serviceID,
-            taskFilter: result.task || '*',
-            outputFilter: result.output || '*'
+            taskFilter: result.taskKey || result.task || '*',
+            outputFilter: result.outputKey || result.output || '*'
         });
         stream.on('data', async ({ outputKey, outputData }) => {
-            await this.executeTask(task, outputKey, outputData);
+            if (result.filter(outputKey, JSON.parse(outputData))) {
+                await this.executeTask(task, outputKey, outputData);
+            }
         });
         return stream;
     }
 
     private executeTask(task: Task, key: string, data: any): Promise<ExecuteTaskReply | Error> {
         return new Promise<ExecuteTaskReply | Error>((resolve, reject) => {
-            const inputData = typeof task.inputs === 'function'
+            const inputData = typeof task.inputs == 'function'
                 ? task.inputs(key, JSON.parse(data))
                 : task.inputs || {};
 
@@ -78,20 +92,54 @@ class Application {
 }
 
 type Event = {
+    // serviceID is service's ID.
     serviceID: string
-    filter?: string
+    
+    // eventKey is event key filter.
+    eventKey?: string
+
+    // filter callback func is used to filter events by event key and
+    // event data before continuing to execute the task.
+    // task execution only will be made when filter returned with a true.
+    // TODO: deprecate string | in future.
+    filter?: string | ((eventKey: string, eventData: Object) => boolean)
 }
 
 type Result = {
+    // serviceID is service's ID.
     serviceID: string
+
+    // DEPRECATED: Please use taskKey instead
     task?: string
+
+    // taskKey is task key filter.
+    taskKey?: string
+
+    // DEPRECATED: Please use outputKey instead
     output?: string
+
+    // outputKey is output key filter.
+    outputKey?: string
+
+    // filter callback func is used to filter task results by output key and
+    // output data before continuing to execute the task.
+    // task execution only will be made when filter returned with a true.
+    filter?: (outputKey: string, outputData: Object) => boolean
 }
 
 type Task = {
+    // serviceID is service's ID.
     serviceID: string
+
+    // taskKey is task's key.
     taskKey: string
-    inputs?: any // TODO(ilgooz): define types
+
+    // inputs is the task's input data.
+    // it can directly get an object as value or a callback func to dynamically
+    // set the inputs depending on relevant event data or task result.
+    // key can be event key or output key.
+    // data can be event data or output data.
+    inputs?: Object | ((key: string, data: Object) => Object)
 }
 
 export default Application;