+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+- Add `filter: (key: string, data: Object) => boolean` on `whenEvent` and `whenResult`.
+This way it's now possible to filter by event key, event data on events and
+output key, output data on results.
+- Normalize the api, use:
+  - `taskKey` instead of `task`.
+  - `outputKey` instead of `output`.
+  - `eventKey` instead of `event`.
+- Update documentation.
+
 ## v1.2.3
 
 Fix `whenResult` callback execution